01-07-HOME SEG_FAULT at paint_miniplayer FIXED

+++ src/load_images.c
@@ -6,7 +6,7 @@
 /*   By: pgruz11 <pgruz11@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/30 20:31:49 by pgruz11           #+#    #+#             */
-/*   Updated: 2024/06/30 23:12:22 by pgruz11          ###   ########.fr       */
+/*   Updated: 2024/07/01 08:57:21 by pgruz11          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,8 +38,10 @@ void	ft_create_minipmap(t_data *d)
 	d->imgs.mini_src = mlx_new_image(d->game, d->maps.minimap_w,
 		d->maps.minimap_h);
 	d->imgs.mini_view = mlx_new_image(d->game, MINI_W, MINI_H);
-	d->maps.map_scale_x = (float)(d->imgs.mini_src->width / d->maps.pix_width);
-	d->maps.map_scale_y = (float)(d->imgs.mini_src->height / d->maps.pix_height);
+	d->maps.map_scale_x = (float)d->imgs.mini_src->width
+		/ (float)d->maps.pix_width;
+	d->maps.map_scale_y = (float)d->imgs.mini_src->height
+		/ (float)d->maps.pix_height;
 	ft_paint_minimap(d, d->maps.minimap_w, d->maps.minimap_h);
 	mlx_image_to_window(d->game, d->imgs.mini_src, 0, 0);
 }
